build sub processes inside CommandPipeline
